split xml writer tests in 1d and 2d test file.
