+++ test/CMakeLists.txt
@@ -39,7 +39,8 @@ set(TEST_SOURCES
         irit_io_2_d_test.cc
         irit_io_3_d_test.cc
         xml_reader_test.cc
-        xml_writer_test.cc)
+        xml_writer_1d_test.cc
+        xml_writer_2d_test.cc)
 
 find_package(GMock REQUIRED)
 #include_directories(${GTEST_INCLUDE_DIRS})
